requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
click==8.1.7
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.3
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pyzmq==25.1.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
uvicorn==0.27.1
wcwidth==0.2.13
zipp==3.17.0
